Decouple creation of DecoderBlock weights/state from golden tensors
