+++ models/demos/deepseek_v3_b1/demo/decoder_stage.py
@@ -12,6 +12,8 @@ import torch
 from loguru import logger
 
 import ttnn
+from models.demos.deepseek_v3.reference.modeling_deepseek import yarn_get_mscale
+from models.demos.deepseek_v3.tt.rope import get_cos_sin_matrix, get_rot_transformation_mat
 from models.demos.deepseek_v3_b1.demo.stage import (
     ACTIVATION_FIFO_SIZE,
     ACTIVATION_PAGE_SIZE_BYTES,
@@ -22,12 +24,672 @@ from models.demos.deepseek_v3_b1.demo.stage import (
 from models.demos.deepseek_v3_b1.fused_ops.attention_block.op import AttentionBlock
 from models.demos.deepseek_v3_b1.fused_ops.decoder_block.op import DecoderBlock
 from models.demos.deepseek_v3_b1.fused_ops.moe.op import MoeOp
+from models.demos.deepseek_v3_b1.micro_ops.flash_mla.op import FlashMLADecode
 from models.demos.deepseek_v3_b1.micro_ops.pipeline_block.op import PipelineBlock
-
-# TODO: This shouldn't live in the test file; we should refactor this
-from models.demos.deepseek_v3_b1.tests.unit_tests.test_decoder_block import create_decoder_block_tensors
+from models.demos.deepseek_v3_b1.micro_ops.sdpa_reduce_to_all.op import compute_forwarder_scratch_size
+from models.demos.deepseek_v3_b1.tests.unit_tests.test_moe_mlp import RoutedExpert, SharedExpert
+from models.demos.deepseek_v3_b1.tests.unit_tests.test_pre_sdpa import deinterleave_kv_cache
 from models.demos.deepseek_v3_b1.utils import get_pinned_optimal_dram_bank_to_logical_worker_assignment
-from models.demos.deepseek_v3_b1.weights.prepare import DeepSeekV3DenseLayerWeights, DeepSeekV3MoELayerWeights
+from models.demos.deepseek_v3_b1.weights.prepare import (
+    DeepSeekV3DenseLayerWeights,
+    DeepSeekV3MoELayerWeights,
+    create_gate_indices_tensor,
+    prepare_dense_layer_weights,
+    prepare_moe_layer_weights,
+)
+
+
+def create_decoder_block_tensors(
+    submesh,
+    mesh_rows,
+    mesh_cols,
+    sender_row,
+    sender_col,
+    position_id,
+    state_dict,
+    layer_idx,
+    max_seq_len,
+    reduce_root_coord=ttnn.MeshCoordinate(1, 1),
+    *,
+    is_moe: bool = True,
+    num_routed_experts: int = 0,
+    preloaded_weights=None,
+    validate_debug_tensors: bool = False,
+    torch_input=None,
+):
+    """Create all tensors required by DecoderBlock.op().
+
+    Returns a dict with all attention + FFN + shared expert + reduce tensors.
+    Intermediate torch CPU tensors (torch_input, torch_kv_cache, etc.) are
+    included so that callers (e.g. golden-reference builders) can reuse them.
+    """
+    if preloaded_weights is None and state_dict is None:
+        raise ValueError("Either state_dict or preloaded_weights must be provided")
+    torch.manual_seed(0)
+    num_devices = mesh_rows * mesh_cols
+    device_grid_size = submesh.compute_with_storage_grid_size()
+    mesh_mapper = ttnn.ReplicateTensorToMesh(submesh)
+
+    # TODO: Shouldn't hardcode this here
+    class _RopeConfig:
+        qk_rope_head_dim = 64
+        rope_theta = 10000.0
+        rope_scaling = {
+            "factor": 40,
+            "original_max_position_embeddings": 4096,
+            "beta_fast": 32,
+            "beta_slow": 1,
+            "mscale": 1.0,
+            "mscale_all_dim": 1.0,
+        }
+
+    _RopeConfig.max_seq_len = max_seq_len
+
+    # ── Constants for runtime tensors ──
+    QNOPE_HEAD_DIM = 128
+    QROPE_HEAD_DIM = _RopeConfig.qk_rope_head_dim
+    QNOPE_OUT_DIM = 512
+    KNOPE_DIM = 512
+    KROPE_DIM = 64
+
+    M = 1
+    K = 7168
+    output_size = 7168
+    shape = (1, K)
+    q_head_dim = QNOPE_HEAD_DIM + QROPE_HEAD_DIM
+    mscale = yarn_get_mscale(
+        _RopeConfig.rope_scaling["factor"],
+        _RopeConfig.rope_scaling["mscale_all_dim"],
+    )
+    scale = q_head_dim**-0.5 * mscale * mscale
+    kvpe_dim = KNOPE_DIM + KROPE_DIM
+
+    QNOPE_GRID_COLS = 8
+    QROPE_GRID_COLS = 4
+    matmul2_grid_y = 8
+    qrope_num_cores = QROPE_GRID_COLS * matmul2_grid_y
+
+    NUM_SDPA_WORKERS = 8
+    SDPA_L_HEIGHT = 8
+    SDPA_L_WIDTH = 512 * NUM_SDPA_WORKERS
+    SDPA_MS_WIDTH = 32 * NUM_SDPA_WORKERS
+
+    mcast_core_x = device_grid_size.x - 1
+    mcast_core_y = 9
+    mcast_core = ttnn.CoreCoord(mcast_core_x, mcast_core_y)
+    tile = ttnn.Tile([1, 32])
+
+    kv_cache_branch_start_offset = (0, 8)
+    kv_cache_branch_rope_crs = ttnn.CoreRangeSet(
+        {
+            ttnn.CoreRange(
+                ttnn.CoreCoord(8 + kv_cache_branch_start_offset[0], kv_cache_branch_start_offset[1]),
+                ttnn.CoreCoord(8 + kv_cache_branch_start_offset[0], 1 + kv_cache_branch_start_offset[1]),
+            )
+        }
+    )
+
+    # ── SDPA KV cache buffer ──
+    kv_cache_num_cores_x = device_grid_size.x
+    kv_cache_num_cores_y = device_grid_size.y
+    kv_cache_num_cores = kv_cache_num_cores_x * kv_cache_num_cores_y
+    kv_cache_shard_height = 256
+    kv_cache_shard_spec = ttnn.ShardSpec(
+        ttnn.CoreRangeSet(
+            {ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(kv_cache_num_cores_x - 1, kv_cache_num_cores_y - 1))}
+        ),
+        (kv_cache_shard_height, kvpe_dim),
+        ttnn.ShardOrientation.ROW_MAJOR,
+    )
+    sdpa_kv_cache_buffer = ttnn.from_torch(
+        torch.randn((kv_cache_shard_height * kv_cache_num_cores, kvpe_dim), dtype=torch.bfloat16),
+        dtype=ttnn.bfloat8_b,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=ttnn.MemoryConfig(
+            ttnn.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.BufferType.L1, kv_cache_shard_spec
+        ),
+        mesh_mapper=mesh_mapper,
+    )
+
+    # ── SDPA output intermediate buffer ──
+    sdpa_out_interm_num_cores = device_grid_size.x * device_grid_size.y
+    sdpa_out_interm_num_slots = 5  # MoE needs 36864 bytes/shard; 5 slots × 17 tiles × 512 = 43520
+    sdpa_out_interm_shard_height = sdpa_out_interm_num_slots * 8
+    sdpa_out_interm_shard_width = 17 * 32
+    sdpa_out_interm_total_height = sdpa_out_interm_shard_height * sdpa_out_interm_num_cores
+    sdpa_out_interm_shard_spec = ttnn.ShardSpec(
+        ttnn.CoreRangeSet(
+            {ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(device_grid_size.x - 1, device_grid_size.y - 1))}
+        ),
+        (sdpa_out_interm_shard_height, sdpa_out_interm_shard_width),
+        ttnn.ShardOrientation.ROW_MAJOR,
+    )
+    sdpa_out_interm_buffer = ttnn.from_torch(
+        torch.zeros((sdpa_out_interm_total_height, sdpa_out_interm_shard_width), dtype=torch.bfloat16),
+        dtype=ttnn.bfloat16,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=ttnn.MemoryConfig(
+            ttnn.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.BufferType.L1, sdpa_out_interm_shard_spec
+        ),
+        mesh_mapper=mesh_mapper,
+        tile=ttnn.Tile([8, 32]),
+    )
+
+    if torch_input is None:
+        torch_input = torch.randn(shape, dtype=torch.bfloat16)
+
+    # ══════════════════════════════════════════════════════════════════════════
+    # All weights via prepare_* functions or preloaded
+    # ══════════════════════════════════════════════════════════════════════════
+    if preloaded_weights is not None:
+        layer = preloaded_weights
+    else:
+        if is_moe:
+            layer = prepare_moe_layer_weights(
+                submesh,
+                state_dict,
+                layer_idx,
+                num_routed_experts=num_routed_experts,
+                move_to_device=True,
+            )
+        else:
+            layer = prepare_dense_layer_weights(submesh, state_dict, layer_idx, move_to_device=True)
+
+    # ── FFN final output config (DRAM streaming matmul output grid) ──
+    gate_proj_noc = ttnn.NOC.NOC_0
+    gate_proj_worker_cores = get_pinned_optimal_dram_bank_to_logical_worker_assignment(submesh, gate_proj_noc)
+    gate_proj_core_ranges = ttnn.CoreRangeSet([ttnn.CoreRange(core, core) for core in gate_proj_worker_cores])
+    num_gate_proj_cores = len(gate_proj_worker_cores)
+    final_output_width_per_core = RoutedExpert.FINAL_OUTPUT_WIDTH_PER_CORE
+    final_output_total_width = final_output_width_per_core * num_gate_proj_cores
+    num_banks = submesh.dram_grid_size().x
+    tile_w = RoutedExpert.TILE_W
+    down_proj_N_padded = ((K + num_banks * tile_w - 1) // (num_banks * tile_w)) * (num_banks * tile_w)
+    per_core_down_proj_N = down_proj_N_padded // num_banks
+
+    final_output_shard_spec = ttnn.ShardSpec(
+        gate_proj_core_ranges,
+        (1, final_output_width_per_core),
+        ttnn.ShardOrientation.ROW_MAJOR,
+    )
+    final_output_mem_config = ttnn.MemoryConfig(
+        ttnn.TensorMemoryLayout.WIDTH_SHARDED, ttnn.BufferType.L1, final_output_shard_spec
+    )
+
+    # ── MoE-only: gate indices and output buffers ──
+    if is_moe:
+        input_core = ttnn.CoreCoord(device_grid_size.x - 1, RoutedExpert.INPUT_CORE_Y)
+        input_core_grid = ttnn.CoreRangeSet([ttnn.CoreRange(input_core, input_core)])
+
+        ttnn_gate_indices = create_gate_indices_tensor(submesh, input_core_grid, mesh_mapper=mesh_mapper)
+
+        tile_1x16 = ttnn.Tile((1, 16))
+        gate_output_shard_spec = ttnn.ShardSpec(input_core_grid, (1, 16), ttnn.ShardOrientation.ROW_MAJOR)
+        gate_output_mem_config = ttnn.MemoryConfig(
+            ttnn.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.BufferType.L1, gate_output_shard_spec
+        )
+        gate_output_scores_tensor = ttnn.from_torch(
+            torch.zeros((1, 16), dtype=torch.bfloat16),
+            dtype=ttnn.bfloat16,
+            layout=ttnn.TILE_LAYOUT,
+            device=submesh,
+            memory_config=gate_output_mem_config,
+            tile=tile_1x16,
+            mesh_mapper=mesh_mapper,
+        )
+        gate_output_indices_tensor = ttnn.from_torch(
+            torch.zeros((1, 16), dtype=torch.uint16),
+            dtype=ttnn.uint16,
+            layout=ttnn.TILE_LAYOUT,
+            device=submesh,
+            memory_config=gate_output_mem_config,
+            tile=tile_1x16,
+            mesh_mapper=mesh_mapper,
+        )
+        moe_ref_gate_output_scores = None
+        moe_ref_gate_output_indices = None
+        if validate_debug_tensors:
+            moe_ref_gate_output_scores = ttnn.from_torch(
+                torch.zeros((1, 16), dtype=torch.bfloat16),
+                dtype=ttnn.bfloat16,
+                layout=ttnn.TILE_LAYOUT,
+                device=submesh,
+                memory_config=gate_output_mem_config,
+                tile=tile_1x16,
+                mesh_mapper=mesh_mapper,
+            )
+            moe_ref_gate_output_indices = ttnn.from_torch(
+                torch.zeros((1, 16), dtype=torch.uint16),
+                dtype=ttnn.uint16,
+                layout=ttnn.TILE_LAYOUT,
+                device=submesh,
+                memory_config=gate_output_mem_config,
+                tile=tile_1x16,
+                mesh_mapper=mesh_mapper,
+            )
+
+    # ══════════════════════════════════════════════════════════════════════════
+    # Attention input/intermediate/output mesh tensors
+    # ══════════════════════════════════════════════════════════════════════════
+    sender_coord = ttnn.MeshCoordinate(sender_row, sender_col)
+    shard_spec = ttnn.ShardSpec(
+        ttnn.CoreRangeSet({ttnn.CoreRange(mcast_core, mcast_core)}), shape, ttnn.ShardOrientation.ROW_MAJOR
+    )
+    mem_config = ttnn.MemoryConfig(ttnn.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.BufferType.L1, shard_spec)
+
+    device_tensors = []
+    for row in range(mesh_rows):
+        for col in range(mesh_cols):
+            if row == sender_row and col == sender_col:
+                device_tensors.append(torch_input)
+            else:
+                device_tensors.append(torch.zeros_like(torch_input))
+
+    input_tensor_mesh = ttnn.from_torch(
+        torch.cat(device_tensors, dim=0),
+        device=submesh,
+        layout=ttnn.TILE_LAYOUT,
+        tile=tile,
+        dtype=ttnn.bfloat16,
+        memory_config=mem_config,
+        mesh_mapper=ttnn.ShardTensorToMesh(submesh, dim=0),
+    )
+
+    # ── RoPE TTNN tensors ──
+    qrope_dram_mem = ttnn.MemoryConfig(ttnn.TensorMemoryLayout.INTERLEAVED, ttnn.BufferType.DRAM)
+    position_ids = torch.tensor([position_id])
+
+    cos_sin_4d, sin_sin_4d = get_cos_sin_matrix(_RopeConfig)
+    torch_cos = cos_sin_4d.squeeze(0).squeeze(0)  # [max_seq_len, dim]
+    torch_sin = sin_sin_4d.squeeze(0).squeeze(0)  # [max_seq_len, dim]
+    torch_trans_mat = get_rot_transformation_mat()
+
+    ttnn_qrope_cos = ttnn.from_torch(
+        torch_cos.unsqueeze(0).unsqueeze(0),
+        dtype=ttnn.bfloat16,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=qrope_dram_mem,
+        tile=tile,
+        mesh_mapper=mesh_mapper,
+    )
+    ttnn_qrope_sin = ttnn.from_torch(
+        torch_sin.unsqueeze(0).unsqueeze(0),
+        dtype=ttnn.bfloat16,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=qrope_dram_mem,
+        tile=tile,
+        mesh_mapper=mesh_mapper,
+    )
+    ttnn_krope_cos = ttnn.from_torch(
+        torch_cos.unsqueeze(0).unsqueeze(0),
+        dtype=ttnn.bfloat16,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=qrope_dram_mem,
+        tile=tile,
+        mesh_mapper=mesh_mapper,
+    )
+    ttnn_krope_sin = ttnn.from_torch(
+        torch_sin.unsqueeze(0).unsqueeze(0),
+        dtype=ttnn.bfloat16,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=qrope_dram_mem,
+        tile=tile,
+        mesh_mapper=mesh_mapper,
+    )
+
+    # ── Trans mat ──
+    qrope_grid = ttnn.CoreRange(
+        ttnn.CoreCoord(QNOPE_GRID_COLS, 0), ttnn.CoreCoord(QNOPE_GRID_COLS + QROPE_GRID_COLS - 1, matmul2_grid_y - 1)
+    )
+    trans_mat_crs = kv_cache_branch_rope_crs.merge(ttnn.CoreRangeSet({qrope_grid}))
+    trans_tile = ttnn.Tile((ttnn.TILE_SIZE, ttnn.TILE_SIZE))
+    trans_shard_spec = ttnn.ShardSpec(trans_mat_crs, (ttnn.TILE_SIZE, ttnn.TILE_SIZE), ttnn.ShardOrientation.ROW_MAJOR)
+    trans_mem = ttnn.MemoryConfig(ttnn.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.BufferType.L1, trans_shard_spec)
+    trans_mat_replicated = torch_trans_mat.repeat(1, 1, qrope_num_cores + kv_cache_branch_rope_crs.num_cores(), 1)
+    ttnn_trans_mat = ttnn.from_torch(
+        trans_mat_replicated,
+        dtype=ttnn.bfloat8_b,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=trans_mem,
+        tile=trans_tile,
+        mesh_mapper=mesh_mapper,
+    )
+
+    # ── Position IDs ──
+    pos_core_grid = ttnn.CoreRangeSet(
+        [ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(device_grid_size.x - 1, device_grid_size.y - 1))]
+    )
+    pos_mem = ttnn.MemoryConfig(
+        ttnn.TensorMemoryLayout.HEIGHT_SHARDED,
+        ttnn.BufferType.L1,
+        ttnn.ShardSpec(pos_core_grid, (1, 1), ttnn.ShardOrientation.ROW_MAJOR),
+    )
+    ttnn_position_ids = ttnn.from_torch(
+        torch.full((device_grid_size.x * device_grid_size.y, 1), position_id, dtype=torch.int32),
+        dtype=ttnn.int32,
+        layout=ttnn.ROW_MAJOR_LAYOUT,
+        device=submesh,
+        memory_config=pos_mem,
+        mesh_mapper=mesh_mapper,
+    )
+
+    # ── KV Cache (ND sharded DRAM) ──
+    program_config = FlashMLADecode.ProgramConfig(k_chunk_size=128, exp_approx_mode=False)
+    grid = program_config.grid
+    kv_nd_shard_spec = ttnn.NdShardSpec(
+        shard_shape=[1, 1, program_config.k_chunk_size, kvpe_dim],
+        grid=grid.optimal_dram_grid(),
+        orientation=ttnn.ShardOrientation.ROW_MAJOR,
+        shard_distribution_strategy=ttnn.ShardDistributionStrategy.ROUND_ROBIN_1D,
+    )
+    kv_mem = ttnn.MemoryConfig(buffer_type=ttnn.BufferType.DRAM, nd_shard_spec=kv_nd_shard_spec)
+    num_sp = mesh_rows
+    dcs = program_config.device_chunk_size
+    torch_kv_cache = torch.zeros((1, 1, max_seq_len, kvpe_dim), dtype=torch.bfloat16)
+    torch_kv_cache[:, :, :position_id, :] = torch.randn(1, 1, position_id, kvpe_dim, dtype=torch.bfloat16)
+    torch_kv_cache_shuffled = deinterleave_kv_cache(torch_kv_cache, dcs, num_sp)
+    kv_cache_2d_mesh_mapper = ttnn.ShardTensor2dMesh(submesh, mesh_shape=(mesh_rows, mesh_cols), dims=(2, None))
+    ttnn_kv_cache = ttnn.from_torch(
+        torch_kv_cache_shuffled,
+        dtype=ttnn.bfloat8_b,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=kv_mem,
+        mesh_mapper=kv_cache_2d_mesh_mapper,
+    )
+
+    # ── KV cache clone for standalone AttentionBlock.op sanity check ──
+    ttnn_kv_cache_attn_ref = None
+    if validate_debug_tensors:
+        ttnn_kv_cache_attn_ref = ttnn.from_torch(
+            torch_kv_cache_shuffled,
+            dtype=ttnn.bfloat8_b,
+            layout=ttnn.TILE_LAYOUT,
+            device=submesh,
+            memory_config=kv_mem,
+            mesh_mapper=kv_cache_2d_mesh_mapper,
+        )
+
+    # ── SDPA output tensor ──
+    s1_cores, _ = FlashMLADecode.ProgramConfig.grid.BLOCKS[0]
+    sdpa_input_output_grid_crs = ttnn.CoreRangeSet(
+        [ttnn.CoreRange(ttnn.CoreCoord(x, y), ttnn.CoreCoord(x, y)) for x, y in s1_cores]
+    )
+    HEADS_PER_ROW = 8
+    SDPA_INPUT_NUM_CORES = len(s1_cores)
+    sdpa_tile = ttnn.Tile([8, 32])
+    sdpa_input_output_shard_spec = ttnn.ShardSpec(
+        sdpa_input_output_grid_crs, (HEADS_PER_ROW, QNOPE_OUT_DIM), ttnn.ShardOrientation.ROW_MAJOR
+    )
+    sdpa_mem = ttnn.MemoryConfig(
+        ttnn.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.BufferType.L1, sdpa_input_output_shard_spec
+    )
+    ttnn_sdpa_output = None
+    if validate_debug_tensors:
+        ttnn_sdpa_output = ttnn.from_torch(
+            torch.zeros((SDPA_INPUT_NUM_CORES * HEADS_PER_ROW, QNOPE_OUT_DIM), dtype=torch.bfloat16),
+            dtype=ttnn.bfloat16,
+            layout=ttnn.TILE_LAYOUT,
+            device=submesh,
+            memory_config=sdpa_mem,
+            mesh_mapper=mesh_mapper,
+            tile=sdpa_tile,
+        )
+
+    # ── Post-SDPA tensors ──
+    a_tile = ttnn.Tile([M, 32])
+    shard_mesh_mapper = ttnn.ShardTensorToMesh(submesh, dim=0)
+    gather_core = ttnn.CoreCoord(12, 9)
+    gather_core_grid = ttnn.CoreRangeSet([ttnn.CoreRange(gather_core, gather_core)])
+
+    # ── Attention block output / MoE residual input (overlapped with sdpa_kv_cache_buffer) ──
+    # These are temporally disjoint: the kv cache on core (12,9) is done after SDPA,
+    # so the attention output and MoE residual input can reuse that L1 region.
+    a_tile_size = a_tile.get_tile_size(ttnn.bfloat16)  # 1×32 tile → 64 bytes
+    num_output_tiles = output_size // 32  # 7168 / 32 = 224 tiles
+    output_shard_spec = ttnn.ShardSpec(gather_core_grid, (M, output_size), ttnn.ShardOrientation.ROW_MAJOR)
+    output_mem_config = ttnn.MemoryConfig(ttnn.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.BufferType.L1, output_shard_spec)
+    mesh_output_torch = torch.cat([torch.zeros((M, output_size), dtype=torch.bfloat16)] * num_devices, dim=0)
+    attn_output = ttnn.from_torch(
+        mesh_output_torch,
+        device=submesh,
+        layout=ttnn.TILE_LAYOUT,
+        tile=a_tile,
+        dtype=ttnn.bfloat16,
+        memory_config=output_mem_config,
+        mesh_mapper=shard_mesh_mapper,
+    )
+    attn_ref_output = None
+    if validate_debug_tensors:
+        attn_ref_output = ttnn.from_torch(
+            mesh_output_torch,
+            device=submesh,
+            layout=ttnn.TILE_LAYOUT,
+            tile=a_tile,
+            dtype=ttnn.bfloat16,
+            memory_config=output_mem_config,
+            mesh_mapper=shard_mesh_mapper,
+        )
+
+    # ── SDPA worker/forwarder tensors ──
+    sdpa_output_cores = FlashMLADecode.ProgramConfig.grid.output_cores(0, NUM_SDPA_WORKERS)
+    sdpa_worker_grid = ttnn.CoreRangeSet(
+        [ttnn.CoreRange(ttnn.CoreCoord(x, y), ttnn.CoreCoord(x, y)) for x, y in sdpa_output_cores]
+    )
+    sdpa_l_per_worker = SDPA_L_WIDTH // NUM_SDPA_WORKERS
+    sdpa_ms_per_worker = SDPA_MS_WIDTH // NUM_SDPA_WORKERS
+
+    sdpa_recv_per_worker = sdpa_l_per_worker + sdpa_ms_per_worker
+    sdpa_recv_shard_shape = (2 * SDPA_L_HEIGHT, sdpa_recv_per_worker)
+    sdpa_recv_mem = ttnn.MemoryConfig(
+        ttnn.TensorMemoryLayout.WIDTH_SHARDED,
+        ttnn.BufferType.L1,
+        ttnn.ShardSpec(sdpa_worker_grid, sdpa_recv_shard_shape, ttnn.ShardOrientation.ROW_MAJOR),
+    )
+    sdpa_recv_full_width = sdpa_recv_per_worker * NUM_SDPA_WORKERS
+    mesh_recv = torch.cat(
+        [torch.zeros((2 * SDPA_L_HEIGHT, sdpa_recv_full_width), dtype=torch.bfloat16)] * num_devices, dim=0
+    )
+    ttnn_sdpa_intermediate_recv = ttnn.from_torch(
+        mesh_recv,
+        device=submesh,
+        dtype=ttnn.bfloat16,
+        layout=ttnn.TILE_LAYOUT,
+        memory_config=sdpa_recv_mem,
+        tile=sdpa_tile,
+        mesh_mapper=shard_mesh_mapper,
+    )
+
+    sdpa_forwarder_cores = [ttnn.CoreCoord(9, 8), ttnn.CoreCoord(10, 8)]
+    sdpa_forwarder_grid = ttnn.CoreRangeSet([ttnn.CoreRange(c, c) for c in sdpa_forwarder_cores])
+    sdpa_fwd_buffer_bytes = compute_forwarder_scratch_size(
+        batch_size=SDPA_L_HEIGHT,
+        l_width=sdpa_l_per_worker,
+        num_cores=NUM_SDPA_WORKERS,
+    )
+    sdpa_fwd_total_elements = sdpa_fwd_buffer_bytes // 2
+    # THIS BUFFER SIZE IS NOT CORRECT BECAUSE WE'RE INCORRECTLY DIVIDING BY 2
+    # TODO: Plan to remove this scratch buffer entirely once we reduce cb memory usage currently being overlapped with this buffer.
+    sdpa_fwd_per_forwarder = sdpa_fwd_total_elements // 2
+    sdpa_forwarder_mem = ttnn.MemoryConfig(
+        ttnn.TensorMemoryLayout.WIDTH_SHARDED,
+        ttnn.BufferType.L1,
+        ttnn.ShardSpec(sdpa_forwarder_grid, (1, sdpa_fwd_per_forwarder), ttnn.ShardOrientation.ROW_MAJOR),
+    )
+    mesh_fwd_scratch = torch.cat([torch.zeros((1, sdpa_fwd_total_elements), dtype=torch.bfloat16)] * num_devices, dim=0)
+    ttnn_sdpa_forwarder_scratch = ttnn.from_torch(
+        mesh_fwd_scratch,
+        device=submesh,
+        dtype=ttnn.bfloat16,
+        layout=ttnn.ROW_MAJOR_LAYOUT,
+        memory_config=sdpa_forwarder_mem,
+        mesh_mapper=shard_mesh_mapper,
+    )
+
+    # ══════════════════════════════════════════════════════════════════════════
+    # Reduce-to-one tensors
+    # ══════════════════════════════════════════════════════════════════════════
+    reduce_mesh_mapper_config = ttnn.MeshMapperConfig([ttnn.PlacementShard(0), ttnn.PlacementShard(1)], submesh.shape)
+    reduce_mesh_mapper = ttnn.create_mesh_mapper(submesh, reduce_mesh_mapper_config)
+    tile_1x32 = ttnn.Tile([1, 32])
+
+    # Single intermediate tensor with 3x shard width for all 3 reduction rounds
+    orig_shard_spec = final_output_mem_config.shard_spec
+    intermediate_shard_shape = [orig_shard_spec.shape[0], orig_shard_spec.shape[1] * 3]
+    intermediate_mem_config = ttnn.MemoryConfig(
+        ttnn.TensorMemoryLayout.WIDTH_SHARDED,
+        ttnn.BufferType.L1,
+        ttnn.ShardSpec(
+            orig_shard_spec.grid,
+            intermediate_shard_shape,
+            ttnn.ShardOrientation.ROW_MAJOR,
+        ),
+    )
+    intermediate_tensors = ttnn.from_torch(
+        torch.zeros([4, 2, final_output_total_width * 3], dtype=torch.bfloat16),
+        dtype=ttnn.bfloat16,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=intermediate_mem_config,
+        tile=tile_1x32,
+        mesh_mapper=reduce_mesh_mapper,
+    )
+
+    shard_cores_list = ttnn.corerange_to_cores(gate_proj_core_ranges, row_wise=True)
+    aggregator_core = shard_cores_list[0]
+    reduce_output_shard_grid = ttnn.CoreRangeSet({ttnn.CoreRange(aggregator_core, aggregator_core)})
+    reduce_output_mem = ttnn.MemoryConfig(
+        ttnn.TensorMemoryLayout.WIDTH_SHARDED,
+        ttnn.BufferType.L1,
+        ttnn.ShardSpec(reduce_output_shard_grid, (1, final_output_total_width), ttnn.ShardOrientation.ROW_MAJOR),
+    )
+    reduce_output_tensor = ttnn.from_torch(
+        torch.zeros([4, 2, final_output_total_width], dtype=torch.bfloat16),
+        dtype=ttnn.bfloat16,
+        layout=ttnn.TILE_LAYOUT,
+        device=submesh,
+        memory_config=reduce_output_mem,
+        tile=tile_1x32,
+        mesh_mapper=reduce_mesh_mapper,
+    )
+
+    # ── Standalone MoE ref reduce tensors (MoE only) ──
+    if is_moe:
+        moe_ref_reduce_intermediate = None
+        moe_ref_reduce_output = None
+        if validate_debug_tensors:
+            moe_ref_reduce_intermediate = ttnn.from_torch(
+                torch.zeros([4, 2, final_output_total_width * 3], dtype=torch.bfloat16),
+                dtype=ttnn.bfloat16,
+                layout=ttnn.TILE_LAYOUT,
+                device=submesh,
+                memory_config=intermediate_mem_config,
+                tile=tile_1x32,
+                mesh_mapper=reduce_mesh_mapper,
+            )
+            moe_ref_reduce_output = ttnn.from_torch(
+                torch.zeros([4, 2, final_output_total_width], dtype=torch.bfloat16),
+                dtype=ttnn.bfloat16,
+                layout=ttnn.TILE_LAYOUT,
+                device=submesh,
+                memory_config=reduce_output_mem,
+                tile=tile_1x32,
+                mesh_mapper=reduce_mesh_mapper,
+            )
+
+    sender_core_from_residual = attn_output.memory_config().shard_spec.grid.bounding_box().end
+    mcast_grid = ttnn.CoreRangeSet([ttnn.CoreRange(ttnn.CoreCoord(0, 0), sender_core_from_residual)])
+
+    # ── Routed weight tensors differ between MoE (list) and dense (single tensor) ──
+    routed_gate = layer.routed_gate_proj[0] if is_moe else layer.routed_gate_proj
+    routed_up = layer.routed_up_proj[0] if is_moe else layer.routed_up_proj
+    routed_down = layer.routed_down_proj[0] if is_moe else layer.routed_down_proj
+
+    result = {
+        # Attention weights (from prepare_*_layer_weights)
+        "gamma_overlapped": layer.attn_norm,
+        "matmul_weights_overlapped": layer.q_a_proj,
+        "rmsnorm2_gamma_overlapped": layer.q_norm,
+        "matmul2_weights_overlapped": layer.q_b_proj,
+        "matmul3_weights_overlapped": layer.kv_b1_proj,
+        "dkv_matmul_weights_overlapped": layer.kv_a_proj,
+        "dkv_rmsnorm_gamma_overlapped": layer.kv_norm,
+        "kv_b2_overlapped": layer.kv_b2_proj,
+        "o_proj_overlapped": layer.o_proj,
+        "ffn_norm_overlapped": layer.ffn_norm,
+        # Attention activation/buffer tensors
+        "input_tensor_mesh": input_tensor_mesh,
+        "ttnn_qrope_sin": ttnn_qrope_sin,
+        "ttnn_qrope_cos": ttnn_qrope_cos,
+        "ttnn_trans_mat": ttnn_trans_mat,
+        "ttnn_krope_cos": ttnn_krope_cos,
+        "ttnn_krope_sin": ttnn_krope_sin,
+        "ttnn_kv_cache": ttnn_kv_cache,
+        "ttnn_kv_cache_attn_ref": ttnn_kv_cache_attn_ref,
+        "ttnn_position_ids": ttnn_position_ids,
+        "scale": scale,
+        "sdpa_kv_cache_buffer": sdpa_kv_cache_buffer,
+        "sdpa_out_interm_buffer": sdpa_out_interm_buffer,
+        "ttnn_sdpa_output": ttnn_sdpa_output,
+        "sender_coord": sender_coord,
+        "ttnn_sdpa_input_l": None,
+        "ttnn_sdpa_input_ms": None,
+        "ttnn_sdpa_output_l": None,
+        "ttnn_sdpa_intermediate_recv": ttnn_sdpa_intermediate_recv,
+        "ttnn_sdpa_forwarder_scratch": ttnn_sdpa_forwarder_scratch,
+        "device_chunk_size": program_config.device_chunk_size,
+        "ttnn_attention_block_output": attn_output,
+        "ttnn_attn_ref_output": attn_ref_output,
+        # FFN tensors (attn_output IS the FFN residual input — overlapped with kv cache)
+        "ttnn_residual_mcast_src": attn_output,
+        "gate_proj_weights": routed_gate,
+        "up_proj_weights": routed_up,
+        "down_proj_weights": routed_down,
+        "final_output_mem_config": final_output_mem_config,
+        "final_output_total_width": final_output_total_width,
+        # Shared expert weights
+        "shared_gate_weights_overlapped": layer.shared_gate_proj,
+        "shared_up_weights_overlapped": layer.shared_up_proj,
+        "shared_down_weights_tensor": layer.shared_down_proj,
+        "shared_k_parallel": SharedExpert.K_PARALLEL,
+        "shared_n_parallel": SharedExpert.N_PARALLEL,
+        # Reduce-to-one
+        "reduce_intermediate_tensors": intermediate_tensors,
+        "reduce_output_tensor": reduce_output_tensor,
+        "reduce_root_coord": reduce_root_coord,
+        "num_gate_proj_cores": num_gate_proj_cores,
+        "per_core_down_proj_N": per_core_down_proj_N,
+        "mcast_grid": mcast_grid,
+        # Intermediate CPU tensors (for golden-reference builders)
+        "torch_input": torch_input,
+        "torch_kv_cache": torch_kv_cache,
+        "torch_sin": torch_sin,
+        "torch_cos": torch_cos,
+        "torch_position_ids": position_ids,
+    }
+    # MoE-only keys
+    if is_moe:
+        result.update(
+            {
+                "gate_mm_overlapped": layer.gate_mm,
+                "ttnn_gate_bias": layer.gate_bias,
+                "ttnn_gate_indices": ttnn_gate_indices,
+                "gate_output_scores_tensor": gate_output_scores_tensor,
+                "gate_output_indices_tensor": gate_output_indices_tensor,
+                "moe_ref_gate_output_scores": moe_ref_gate_output_scores,
+                "moe_ref_gate_output_indices": moe_ref_gate_output_indices,
+                "moe_ref_reduce_intermediate": moe_ref_reduce_intermediate,
+                "moe_ref_reduce_output": moe_ref_reduce_output,
+            }
+        )
+    return result
 
 
 class DecoderStage(StageKind):
